faster reset in pybullet
